Added internal uprights to link 2 assem
The fillets are rlly aggresive ik, but it's sort of a reminder that it's unfinished. The one thing that needs to be done still is to increase the hole size on all of the bevel gear holes (two features) to accommodate a bearing to fit the bevel gear. I do not remember if we decided which bearing to use so I made them the same size.