Link 1 to Link 2 Motormount/connection
This is the motor mount that will also act as the link 1 to link 2 connection. The bottom plate will screw into the main link 1 shaft (using countersunk screws) and then the motor will go on the L1-L2 bracket. Depending on the gearbox, I imagine the mounting holes will be different, but it should look similar. We use a similar setup at my work, and we just have a 3D printed negative underneath to help with motor sag.

Then, in the indent, there will be a thrust bearing to handle the moment from the weight of link 2.

Thanks