Adjusted motor housing & shaft
Think the shaft and housing are complete-ish (still need to add corner relief in housing)

Biggest thing left is to fix the top lid (make it interface with the top bearing) and make the bottom lid. Have to consider assembly during this as well to maintain axial alignment

Might also adjust the housing so it's more snug on the bottom bearing, back have to think about it.

Also made all the equations off a text document which is pretty cool. Don't know why it didn't default to excel, but it is what it is ig.